Changes to enable Docker deployment
Some changes have been done to enable the Go
app to be started in a Docker container. The
Docker daemon socket will have to be mounted
as a volume for the app to be able to spawn
sibling containers.

It does not work yet since the bindings are
defined by the container and not the host
anymore, so the paths do not match anymore.